code-rev: moving towards modular approach
part 1
